+++ apps/client/src/app/pages/portfolio/analysis/analysis-page.component.ts
@@ -30,9 +30,12 @@ export class AnalysisPageComponent implements OnDestroy, OnInit {
   public dateRangeOptions = ToggleComponent.DEFAULT_DATE_RANGE_OPTIONS;
   public daysInMarket: number;
   public deviceType: string;
+  public dividendsByMonth: InvestmentItem[];
+  public dividendTimelineDataLabel = $localize`Dividend`;
   public firstOrderDate: Date;
   public hasImpersonationId: boolean;
   public investments: InvestmentItem[];
+  public investmentTimelineDataLabel = $localize`Deposit`;
   public investmentsByMonth: InvestmentItem[];
   public isLoadingBenchmarkComparator: boolean;
   public isLoadingInvestmentChart: boolean;
@@ -42,6 +45,7 @@ export class AnalysisPageComponent implements OnDestroy, OnInit {
   ];
   public performanceDataItems: HistoricalDataItem[];
   public performanceDataItemsInPercentage: HistoricalDataItem[];
+  public portfolioEvolutionDataLabel = $localize`Deposit`;
   public top3: Position[];
   public user: User;
 
@@ -161,6 +165,18 @@ export class AnalysisPageComponent implements OnDestroy, OnInit {
         this.isLoadingInvestmentChart = false;
 
         this.updateBenchmarkDataItems();
+
+        this.changeDetectorRef.markForCheck();
+      });
+
+    this.dataService
+      .fetchDividends({
+        groupBy: 'month',
+        range: this.user?.settings?.dateRange
+      })
+      .pipe(takeUntil(this.unsubscribeSubject))
+      .subscribe(({ dividends }) => {
+        this.dividendsByMonth = dividends;
 
         this.changeDetectorRef.markForCheck();
       });